// src/models/Person.ts
export interface Person {
    ardaId?: number;
    ggId: string;
    name: string;
    comparableName?: string;
    username?: string;
    professionalHeadline: string;
    imageUrl?: string;
    completion?: number;
    grammar?: number;
    weight?: number;
    verified?: boolean;
    totalStrength?: number;
    pageRank?: number;
    isSearchable?: boolean;
}